+++ cms/src/routes/index.js
@@ -1,0 +1,3 @@
+export { default as publishRouter } from './publish';
+export { default as unpublishRouter } from './unpublish';
+export { default as imagesRouter } from './images';
