+++ cms/src/routes/publish.js
@@ -1,9 +1,9 @@
 import express from 'express';
-import { indexOneToES, indexUpdatesToES, indexAllToES } from '../services/publish/sync-ES';
+import { indexOneToES, indexUpdatesToES, indexAllToES } from '../services/elastic-search/sync-ES';
 
-export const publish_router = express.Router();
+const publishRouter = express.Router();
 
-publish_router.post('/model/:modelId', async (req, res) => {
+publishRouter.post('/model/:modelId', async (req, res) => {
   const { modelId } = req.params;
   indexOneToES({
     _id: modelId,
@@ -16,7 +16,7 @@ publish_router.post('/model/:modelId', async (req, res) => {
     );
 });
 
-publish_router.post('/updated', async (req, res) => {
+publishRouter.post('/updated', async (req, res) => {
   indexUpdatesToES()
     .then(data => res.json(data))
     .catch(error =>
@@ -26,7 +26,7 @@ publish_router.post('/updated', async (req, res) => {
     );
 });
 
-publish_router.post('/all', async (req, res) => {
+publishRouter.post('/all', async (req, res) => {
   indexAllToES()
     .then(data => res.json(data))
     .catch(error =>
@@ -35,3 +35,5 @@ publish_router.post('/all', async (req, res) => {
       }),
     );
 });
+
+export default publishRouter;
